Address comment: Add new test to bit-utility-test instead of utility-test, and move bitmap tests from bit_util_test.cc into the new file
