+++ cpp/src/arrow/util/CMakeLists.txt
@@ -49,7 +49,6 @@ add_arrow_test(utility-test
                SOURCES
                align_util_test.cc
                atfork_test.cc
-               bitmap_test.cc
                byte_size_test.cc
                byte_stream_split_test.cc
                cache_test.cc
@@ -99,8 +98,16 @@ add_arrow_test(bit-utility-test
                SOURCES
                bit_block_counter_test.cc
                bit_util_test.cc
+               bitmap_test.cc
                bpacking_test.cc
-               rle_encoding_test.cc)
+               rle_encoding_test.cc
+               test_common.cc)
+
+add_arrow_test(crc32-test
+               SOURCES
+               crc32_test.cc
+               EXTRA_LINK_LIBS
+               Boost::headers)
 
 add_arrow_test(threading-utility-test
                SOURCES
@@ -111,16 +118,10 @@ add_arrow_test(threading-utility-test
                test_common.cc
                thread_pool_test.cc)
 
-add_arrow_test(crc32-test
-               SOURCES
-               crc32_test.cc
-               EXTRA_LINK_LIBS
-               Boost::headers)
-
 add_arrow_benchmark(bit_block_counter_benchmark)
 add_arrow_benchmark(bit_util_benchmark)
-add_arrow_benchmark(bpacking_benchmark)
 add_arrow_benchmark(bitmap_reader_benchmark)
+add_arrow_benchmark(bpacking_benchmark)
 add_arrow_benchmark(cache_benchmark)
 add_arrow_benchmark(compression_benchmark)
 add_arrow_benchmark(decimal_benchmark)
